v2.5.0.3 Beta

+++ installer/installer.spec
@@ -1,17 +1,23 @@
 # -*- mode: python ; coding: utf-8 -*-
 from pathlib import Path
 import re
+import os
+
+def IsWindows() -> bool:
+    return os.name == 'nt'
 
 block_cipher = None
 console = False
 
 # unfortunately runtime_hooks is not relative to the location of the spec file, but everything else is
-if Path('add_lib_path.py').exists():
-    runtime_hooks=['add_lib_path.py']
-elif Path('installer/add_lib_path.py').exists():
-    runtime_hooks=['installer/add_lib_path.py']
-else:
-    raise Exception('failed to find add_lib_path.py')
+runtime_hooks = []
+if IsWindows():
+    if Path('add_lib_path.py').exists():
+        runtime_hooks=['add_lib_path.py']
+    elif Path('installer/add_lib_path.py').exists():
+        runtime_hooks=['installer/add_lib_path.py']
+    else:
+        raise Exception('failed to find add_lib_path.py')
 
 # we can also create an exe for bingo in the same folder
 a_installer = Analysis(
@@ -114,25 +120,26 @@ coll = COLLECT(
 )
 
 # tidy up
-print('tidying up')
-is_lib = re.compile(r'(\.dll$)|(\.so(\..+)?$)|(\.pyd$)')
+if IsWindows():
+    print('tidying up')
+    is_lib = re.compile(r'(\.dll$)|(\.so(\..+)?$)|(\.pyd$)')
 
-dxr = Path('dist') / 'DXRando'
-if not dxr.exists():
-    dxr = Path('installer') / 'dist' / 'DXRando'
+    dxr = Path('dist') / 'DXRando'
+    if not dxr.exists():
+        dxr = Path('installer') / 'dist' / 'DXRando'
 
-if (dxr / 'installer.app').exists():
-    (dxr / 'installer.app').unlink()
+    if (dxr / 'installer.app').exists():
+        (dxr / 'installer.app').unlink()
 
-(dxr / 'lib').mkdir(exist_ok=True)
-for f in dxr.glob('*'):
-    name = f.name.lower()
-    if not is_lib.search(name):
-        continue
-    if name.startswith('python') or name.startswith('libpython'):
-        continue# we need the main python library
-    dest = f.parent / 'lib' / f.name
-    print('renaming', f, 'to', dest)
-    f.rename(dest)
+    (dxr / 'lib').mkdir(exist_ok=True)
+    for f in dxr.glob('*'):
+        name = f.name.lower()
+        if not is_lib.search(name):
+            continue
+        if name.startswith('python') or name.startswith('libpython'):
+            continue# we need the main python library
+        dest = f.parent / 'lib' / f.name
+        print('renaming', f, 'to', dest)
+        f.rename(dest)
 
-print('done tidying up')
+    print('done tidying up')
